WIP move AST desugaring to its own module
